Trekker ut retryListener.
